.history/backend/requirements_20241212222356.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pillow
torch
tenserflow
torchvision
